# examples/demo-backend/openapi/complex.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/customers": {
      "get": {
        "operationId": "Customers_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "ada",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List customers using offset pagination and search",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/customers/{id}": {
      "get": {
        "operationId": "Customers_get",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomerResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "Get one customer with nested address",
        "tags": [
          "Customers"
        ]
      }
    },
    "/api/v1/orders": {
      "get": {
        "operationId": "ComplexOrders_list",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "allOf": [
                {
                  "$ref": "#/components/schemas/Object"
                }
              ]
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "example": "complex-order-001",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "draft",
                "awaiting-payment",
                "paid",
                "fulfillment",
                "shipped",
                "cancelled"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexOrdersResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List tenant orders using cursor pagination",
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "ComplexOrders_create",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Repeating a request with the same key returns the original order.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "checkout-6c5f92ad"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateComplexOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexOrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Business or optimistic-lock conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Create an order idempotently",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}": {
      "get": {
        "operationId": "ComplexOrders_get",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexOrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "Get one order and its state",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/orders/{id}/cancel": {
      "post": {
        "operationId": "ComplexOrders_cancel",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplexOrderResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "The requested resource does not exist.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Business or optimistic-lock conflict.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          },
          {
            "csrf": []
          }
        ],
        "summary": "Apply a validated order state transition",
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/v1/payments": {
      "get": {
        "operationId": "Payments_payments",
        "parameters": [
          {
            "name": "X-Demo-Scenario",
            "in": "header",
            "description": "Forces a deterministic frontend-testing scenario for this request.",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "normal",
                "slow",
                "timeout",
                "server-error",
                "rate-limited",
                "empty",
                "expired-auth",
                "forbidden",
                "conflict",
                "large-dataset"
              ]
            }
          },
          {
            "name": "X-Organization-Id",
            "in": "header",
            "description": "Current tenant. The authenticated user must be a member.",
            "required": true,
            "schema": {
              "type": "string",
              "example": "org-acme"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentsResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request or validation error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is missing or expired.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "The current user lacks permission.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "429": {
            "description": "Demo rate limit scenario.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected or simulated server error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "cookieSession": []
          }
        ],
        "summary": "List payments for tenant orders",
        "tags": [
          "Payments and promotions"
        ]
      }
    },
...
  },
  "info": {
    "title": "Demo Complex API",
    "description": "Cookie-session, multitenant and realtime API for large frontend applications.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3002",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "cookieSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "demo_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Required for authenticated mutations."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 404
          },
          "code": {
            "type": "string",
            "example": "PRODUCT_NOT_FOUND"
          },
          "message": {
            "type": "string",
            "example": "Product not found"
          },
          "details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ErrorDetailDto"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time",
            "example": "2026-07-30T12:00:00.000Z"
          },
          "path": {
            "type": "string",
            "example": "/api/v1/products/product-404"
          },
          "requestId": {
            "type": "string",
            "example": "req-5c9f7a3d"
          }
        },
        "required": [
          "statusCode",
          "code",
          "message",
          "details",
          "timestamp",
          "path",
          "requestId"
        ]
      },
...
      "Object": {
        "type": "object",
        "properties": {}
      },
...
      "CustomersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CustomerDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/PageMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
      "CustomerResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/CustomerDto"
          }
        },
        "required": [
          "data"
        ]
      },
...
      "ComplexOrdersResponseDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComplexOrderDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/CursorMetaDto"
          }
        },
        "required": [
          "data",
          "meta"
        ]
      },
...
    }
  }
}